@@ -1,16 +1,16 @@
 // import {Component, Input} from '@angular/core';
 // import {CompanyProfile} from '../app/credit-rating/profile.service';
-// import {Observable} from 'rxjs/Observable';
+// import {Observable} from "rxjs";
 //
 // @Component({
-//   selector: 'app-search',
+//   selector: 'app-search-bar',
 //   template: '',
 // })
 // export class MockSearchComponent  {
 // }
 //
 // @Component({
-//   selector: 'app-profile',
+//   selector: 'app-credit-report',
 //   template: '',
 // })
 // export class MockProfileComponent  {
