@@ -1,6 +1,6 @@
-import { TestBed, async, inject } from '@angular/core/testing';
+import {inject, TestBed} from '@angular/core/testing';
 
-import { AuthenticationGuard } from './authentication.guard';
+import {AuthenticationGuard} from './authentication.guard';
 
 describe('AuthenticationGuard', () => {
   beforeEach(() => {
