@@ -2,7 +2,7 @@ import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 import {MatCardModule} from '@angular/material';
 
 import {DashboardComponent} from './dashboard.component';
-import {MockProfileComponent, MockSearchComponent} from '../../../testing/mock.components.spec';
+// import {MockProfileComponent, MockSearchComponent} from '../../../testing/mock.components.spec';
 
 describe('DashboardComponent', () => {
   let component: DashboardComponent;
@@ -15,8 +15,8 @@ describe('DashboardComponent', () => {
       ],
       declarations: [
         DashboardComponent,
-        MockSearchComponent,
-        MockProfileComponent,
+        // MockSearchComponent,
+        // MockProfileComponent,
       ],
     })
       .compileComponents();
